refactor: merge files into one
Combine multiple files into a single file to reduce unnecessary module
creation. This change simplifies navigation and improves codebase
organization.
