--- src/components/textarea/textarea-remove.js
@@ -1,10 +1,0 @@
-'use strict'
-
-function removeTextarea(config) {
-	const { id } = config
-	const TEXTAREA = document.getElementById(id)
-	if (!TEXTAREA) return
-	TEXTAREA.remove()
-}
-
-export { removeTextarea }
